op-batcher: Move send transaction to txmgr.go
