+++ op-batcher/driver.go
@@ -15,7 +15,6 @@ import (
 	"github.com/ethereum-optimism/optimism/op-proposer/txmgr"
 	"github.com/ethereum/go-ethereum/accounts"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/log"
 	hdwallet "github.com/miguelmota/go-ethereum-hdwallet"
@@ -221,49 +220,6 @@ func (l *BatchSubmitter) loadBlocksIntoState(ctx context.Context) {
 		l.lastSubmittedBlock = prevID
 		// l.log.Info("added L2 block to channel", "block", prevID, "channel_id", l.ch.ID(), "tx_count", len(block.Transactions()), "time", block.Time())
 	}
-}
-
-func (l *BatchSubmitter) submitTransaction(data []byte) error {
-	// Query for the submitter's current nonce.
-	walletAddr := crypto.PubkeyToAddress(l.cfg.PrivKey.PublicKey)
-	ctx, cancel := context.WithTimeout(l.ctx, time.Second*10)
-	nonce, err := l.cfg.L1Client.NonceAt(ctx, walletAddr, nil)
-	cancel()
-	if err != nil {
-		l.log.Error("unable to get current nonce", "err", err)
-		return err
-	}
-
-	// Create the transaction
-	ctx, cancel = context.WithTimeout(l.ctx, time.Second*10)
-	tx, err := l.CraftTx(ctx, data, nonce)
-	cancel()
-	if err != nil {
-		l.log.Error("unable to craft tx", "err", err)
-		return err
-	}
-
-	// Construct the a closure that will update the txn with the current gas prices.
-	updateGasPrice := func(ctx context.Context) (*types.Transaction, error) {
-		l.log.Debug("updating batch tx gas price")
-		return l.UpdateGasPrice(ctx, tx)
-	}
-
-	// Wait until one of our submitted transactions confirms. If no
-	// receipt is received it's likely our gas price was too low.
-	// TODO: does the tx manager nicely replace the tx?
-	//  (submit a new one, that's within the channel timeout, but higher fee than previously submitted tx? Or use a cheap cancel tx?)
-	ctx, cancel = context.WithTimeout(l.ctx, time.Second*time.Duration(l.cfg.ChannelTimeout))
-	receipt, err := l.txMgr.Send(ctx, updateGasPrice, l.cfg.L1Client.SendTransaction)
-	cancel()
-	if err != nil {
-		l.log.Warn("unable to publish tx", "err", err)
-		return err
-	}
-
-	// The transaction was successfully submitted.
-	l.log.Info("tx successfully published", "tx_hash", receipt.TxHash)
-	return nil
 }
 
 func (l *BatchSubmitter) loop() {
